requirements.txt
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.10.0
APScheduler==3.11.0
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
black==25.1.0
bleach==6.2.0
blinker==1.9.0
boto3==1.40.64
botocore==1.40.64
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.3.0
firebase_admin==7.1.0
flake8==7.2.0
Flask==3.0.2
flask-babel==4.0.0
Flask-Limiter==3.13
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
fonttools==4.58.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.38.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.11.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-logging==3.8.0
google-cloud-monitoring==1.1.0
google-cloud-resource-manager==1.14.2
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-genai==1.31.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.1
multi_key_dict==2.0.3
mypy==1.16.0
mypy_extensions==1.1.0
narwhals==1.41.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
opentelemetry-api==1.36.0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pbr==7.0.2
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.11
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.3.2
PyGithub==2.8.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-gitlab==7.0.0
python-jenkins==1.8.3
python-magic-bin==0.4.14
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.14.0
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
streamlit==1.45.1
stripe==13.1.1
tenacity==9.1.2
textblob==0.19.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.1
wrapt==1.17.3
WTForms==3.2.1
zipp==3.23.0
